Move more logic from DataStoreTable into composable
